Don't completely drop changes that have bad mappings
Fixes #284
